Remove grok from opengever.private
