+++ opengever/meeting/browser/meetingdossier_forms.py
@@ -1,0 +1,6 @@
+from opengever.dossier.browser.forms import DossierAddView
+
+
+class MeetingDossierAddView(DossierAddView):
+    """Add view for opengever.meeting.dossier
+    """
