+++ opengever/meeting/dossier.py
@@ -1,6 +1,4 @@
-from five import grok
 from opengever.dossier.base import DossierContainer
-from opengever.dossier.behaviors.dossier import AddForm
 from opengever.meeting.model import Meeting
 
 
@@ -8,7 +6,3 @@ class MeetingDossier(DossierContainer):
 
     def get_meeting(self):
         return Meeting.query.by_dossier(self).first()
-
-
-class MeetingDossierAddForm(AddForm):
-    grok.name('opengever.meeting.meetingdossier')
