+++ opengever/private/configure.zcml
@@ -2,13 +2,10 @@
     xmlns="http://namespaces.zope.org/zope"
     xmlns:genericsetup="http://namespaces.zope.org/genericsetup"
     xmlns:profilehook="http://namespaces.zope.org/profilehook"
-    xmlns:grok="http://namespaces.zope.org/grok"
     xmlns:i18n="http://namespaces.zope.org/i18n"
     i18n_domain="opengever.private">
 
   <i18n:registerTranslations directory="locales" />
-
-  <grok:grok package="." />
 
   <genericsetup:registerProfile
       name="default"
